Pass through directly_visible flag for area light sources.
I encountered problems where derivatives involving lighting seemed to not be computed correctly, and I don't remember the deails, nor did I ever really understand what was going on.  But I did narrow things down to a point where I realized that turning off direct visibility for the area light source in my test scene would avoid the bad behavior that I was seeing.  But I also noticed that I couldn't turn off direct visibility for area lights as I could for other types of lights, even though the flag existed at a lower level.  So this change simply passes a directly_visible flag thru from the external API.  Of course, it would be great to actually fix the bad behavior that I was originally seeing for which this is a workaround, but having this flag is a good thing in any case, I'd think.  I think I could reproduce that original bad behavior, but I don't think I'd have the ability to fix it.  If you'd be interested in having a repro test case, we could talk.