Small but important bug fix for environment map wrapping behavior.
In the envmap_pdf() lookup function, there's a place where a dx and
a dy are computed and these are clearly supposed to be in the 0-1
range.  But the x and y can wrap around, leading to dx or dy being
far outside that range.  I added a correction for that case.